modules/ozmail: Condense email conf into new module #439
